pyproject.toml
[tool.poetry]
name = "django-state-machine-demo"
version = "0.1.0"
description = ""
authors = ["hassanselim0"]
readme = "README.md"
packages = [{include = "django_state_machine_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-model-utils = "^4.2.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.2"
black = {version = "^22.8.0", allow-prereleases = true}
ipython = "^8.5.0"

[tool.black]
line-length = 100

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-class-docstring",
    "missing-module-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
    "unused-argument",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
